Rework is_ets[4|6]_project()
KNXProjContents seems to be a nice place to host the is_ets4_project()
method. This way, we avoid passing the schema version to the method all
the time.
is_ets6_project() should logically also go inside KNXProjContents. However,
the method is only called once and before KNXProjContents gets created.
Therefore, I prefer to remove it and inline the check.