Add ETS4 support (#241)
* Make project_uid optional

In case "Puid" is not available (e.g. ETS4 export), set it to None.

* Use the correct filename for ETS4 export

The project filename in an ETS4 .knxproj is named "Project.xml" whereas
it is called "project.xml" in later version of ETS.

* Add unit tests for ETS4 .knxproj

The tests uses a dummy project created with ETS 4.2.0 (Build 3884). The
same project has been exported with and without a password.

* Add stub for ETS4

Seems like communication object definitions are missing.

* Be more specific about the ETS version

ETS 4.0 was likely schema version 10.

* Move is_ets[46]_project to utils

* Fix resolve_com_object_ref_id for ETS4

* Add support for ETS4 to communication object

The definition of the links in ETS4 is different than in ETS5/6. ETS4
uses sub-elements whereas ETS5/6 uses a string of GA separated by a
space character.

* Update the stub for ETS4

* Save schema version in KNXProjContents

* Parse element based on schema version

Do not try to parse "Links" (ETS5/6) first than "Connectors" (ETS4) and
see if something succeeds. Use the schema version to call the right
parsing strategy.

* Improve __get_links_from_ets4

Also fix a bug that resulted in the receive GA not being parsed
correctly.

* Parse Buildings for ETS4

ETS4 uses "Buildings" and "BuildingPart" instead of "Locations" and
"Space".

* Create instances outside the loops

* Rename __element_name to _element_name

* Avoid joining and spliting strings for Links

* Add ETS4 .knxproj to refresh_stubs.py

* Rework is_ets[4|6]_project()

KNXProjContents seems to be a nice place to host the is_ets4_project()
method. This way, we avoid passing the schema version to the method all
the time.
is_ets6_project() should logically also go inside KNXProjContents. However,
the method is only called once and before KNXProjContents gets created.
Therefore, I prefer to remove it and inline the check.