feat(Deps): Replace crytpography with hashlib
Cryptography uses rust as engine. This can lead to trouble when trying to build this package for
arm/v7 as cryptography as no prebuilt binaries for this platform. In general, this will help with
cross compatibility across architectures. Also it was needed, when i tried to build a docker image
of an app using xknxproject as a lib. cross building for linux/arm/v7 would always fail in this case
due to having to setup a rust compiler and not succeeding to do so.